requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
